requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
openai==1.52.0
psutil==5.9.8
yt-dlp>=2024.12.23
PyNaCl==1.5.0
aiosqlite==0.20.0
